Rename package m72 -> r72
"m" means "milestone"
"r" means "release"

There was a 7.2 GA, but not a 7.2 milestone
