+++ .teamcity/test-buckets.json
@@ -2028,7 +2028,7 @@
 					"org.gradle.plugins.ide.tooling.r30.ToolingApiEclipseModelClasspathContainerCrossVersionSpec=crossVersionTest",
 					"org.gradle.plugins.ide.eclipse.EclipseTestConfigurationsWithProjectDependenciesIntegrationTest=integTest",
 					"org.gradle.plugins.ide.eclipse.EclipseJavaProjectIntegrationTest=integTest",
-					"org.gradle.plugins.ide.tooling.m72.CompositeBuildCrossVersionSpec=crossVersionTest",
+					"org.gradle.plugins.ide.tooling.r72.CompositeBuildCrossVersionSpec=crossVersionTest",
 					"org.gradle.plugins.ide.eclipse.EclipseProjectNameDeduplicationIntegrationTest=integTest",
 					"org.gradle.plugins.ide.tooling.r211.ToolingApiEclipseModelCrossVersionSpec=crossVersionTest",
 					"org.gradle.plugins.ide.idea.IdeaIntegrationTest=integTest",
@@ -2135,7 +2135,7 @@
 					"org.gradle.plugins.ide.tooling.r30.ToolingApiEclipseModelClasspathContainerCrossVersionSpec=crossVersionTest",
 					"org.gradle.plugins.ide.eclipse.EclipseTestConfigurationsWithProjectDependenciesIntegrationTest=integTest",
 					"org.gradle.plugins.ide.eclipse.EclipseJavaProjectIntegrationTest=integTest",
-					"org.gradle.plugins.ide.tooling.m72.CompositeBuildCrossVersionSpec=crossVersionTest",
+					"org.gradle.plugins.ide.tooling.r72.CompositeBuildCrossVersionSpec=crossVersionTest",
 					"org.gradle.plugins.ide.eclipse.EclipseProjectNameDeduplicationIntegrationTest=integTest",
 					"org.gradle.plugins.ide.tooling.r211.ToolingApiEclipseModelCrossVersionSpec=crossVersionTest",
 					"org.gradle.plugins.ide.idea.IdeaIntegrationTest=integTest",
