+++ subprojects/ide/src/crossVersionTest/groovy/org/gradle/plugins/ide/tooling/r72/CompositeBuildCrossVersionSpec.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.plugins.ide.tooling.m72
+package org.gradle.plugins.ide.tooling.r72
 
 import org.gradle.integtests.tooling.fixture.TargetGradleVersion
 import org.gradle.integtests.tooling.fixture.ToolingApiSpecification
